apicompropy/requirements.txt
Pillow
django
djangorestframework
djoser
django-taggit
django-widget-tweaks